+++ src/content/components/tooltip.css
@@ -1,0 +1,60 @@
+.wanderland-popover {
+  width: 400px !important;
+  height: 400px !important;
+  background-color: white;
+  border-radius: 15px;
+  overflow: hidden;
+  box-shadow: 5px 10px 18px #888888;
+}
+
+.wanderland-popover .wl-content {
+  border-bottom-right-radius: 10px !important;
+  border-bottom-left-radius: 10px !important;
+}
+
+.wanderland-popover .wl-navbar {
+  border-top-right-radius: 10px !important;
+  border-top-left-radius: 10px !important;
+  text-align: center;
+  height: 50px;
+  background-color: #ec7c93;
+}
+
+.wanderland-popover .navbar-brand {
+  font-family: 'Arial Narrow Bold', sans-serif;
+  font-size: 30px;
+}
+
+.anchor-like-button {
+  align-items: normal;
+  background-color: rgba(0, 0, 0, 0);
+  border-color: #ec7c93;
+  border-style: none;
+  box-sizing: content-box;
+  color: #ec7c93;
+  cursor: pointer;
+  display: inline;
+  font: inherit;
+  height: auto;
+  padding: 0;
+  perspective-origin: 0 0;
+  text-align: start;
+  text-decoration: underline;
+  transform-origin: 0 0;
+  width: auto;
+  -moz-appearance: none;
+  -webkit-logical-height: 1em; /* Chrome ignores auto, so we have to use this hack to set the correct height  */
+  -webkit-logical-width: auto; /* Chrome ignores auto, but here for completeness */
+}
+
+@font-face {
+  font-family: 'Indie Flower';
+  font-style: normal;
+  font-weight: 400;
+  src: local('Indie Flower'), local('IndieFlower'),
+      url('./IndieFlower.woff2')
+          format('woff2');
+  unicode-range: U+0000-00FF, U+0131, U+0152-0153, U+02BB-02BC, U+02C6, U+02DA, U+02DC,
+      U+2000-206F, U+2074, U+20AC, U+2122, U+2191, U+2193, U+2212, U+2215, U+FEFF,
+      U+FFFD;
+}
