--- src/content/tooltip.css
@@ -1,38 +1,0 @@
-.wanderland-popover {
-  max-width: 405px !important;
-  max-height: 399px !important;
-  background-color: white;
-  border-radius: 5px;
-  font-family: 'Indie Flower';
-}
-
-.wanderland-popover .popover-content {
-  padding: 0px 0px !important;
-}
-
-.wanderland-popover .navbar {
-  border-bottom-right-radius: 0px !important;
-  border-bottom-left-radius: 0px !important;
-}
-
-.anchor-like-button {
-  align-items: normal;
-  background-color: rgba(0, 0, 0, 0);
-  border-color: #ec7c93;
-  border-style: none;
-  box-sizing: content-box;
-  color: #ec7c93;
-  cursor: pointer;
-  display: inline;
-  font: inherit;
-  height: auto;
-  padding: 0;
-  perspective-origin: 0 0;
-  text-align: start;
-  text-decoration: underline;
-  transform-origin: 0 0;
-  width: auto;
-  -moz-appearance: none;
-  -webkit-logical-height: 1em; /* Chrome ignores auto, so we have to use this hack to set the correct height  */
-  -webkit-logical-width: auto; /* Chrome ignores auto, but here for completeness */
-}
