+++ src/content/tooltipInsert.js
@@ -1,52 +1,8 @@
 import findAndReplaceDOMText from 'findandreplacedomtext';
-import React, { Component } from 'react';
+import React from 'react';
 import ReactDOM from 'react-dom';
-import Popover from 'react-popover';
-import TooltipMap from './tooltipMap';
-import TooltipNavbar from './components/tooltipNavbar';
-import './tooltip.css';
+import Tooltip from './components/tooltip';
 
-const PopoverContent = ({ title }) => (
-    <div className="wanderland-popover">
-        <TooltipNavbar />
-        <div className="tooltipdiv">
-            <TooltipMap title={title} />
-        </div>
-    </div>
-);
-
-class Tooltip extends Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            show: false,
-        };
-        this.setShow = this.setShow.bind(this);
-        this.toggleShow = this.toggleShow.bind(this);
-    }
-
-    setShow(state) {
-        this.setState({ show: state });
-    }
-
-    toggleShow() {
-        this.setState(prevState => ({ show: !prevState.show }));
-    }
-
-    render() {
-        return (
-            <Popover
-                isOpen={this.state.show}
-                body={[<PopoverContent title={this.props.place.title} />]}
-                onOuterAction={() => this.setShow(false)}
-            >
-                <button type="button" className="anchor-like-button" onClick={this.toggleShow}>
-                    {this.props.place.title}
-                </button>
-            </Popover>
-        );
-    }
-}
 
 const TAGS_NOT_TO_COVER = ['a', 'span', 'button']; // Tags not to cover for tooltip wrapping
 const NUM_PARENTS_TO_CHECK = 2; // Number of parents of a node to check for tooltip wrapping
